sai-ingestor/requirements.txt
Flask==3.0.3
pynut2==2.1.2
requests
paho-mqtt
prometheus_client